line geometry and line stepper changes, delete examples

+++ src/Renderer/Geom.cpp
@@ -10,23 +10,23 @@ Renderer::Geom::Matrix33f Renderer::Geom::MatrixFactory::e3() {
 Renderer::Geom::Matrix33f Renderer::Geom::MatrixFactory::rotX(float angle) {
 	return Matrix33f{
 			1, 0 , 0,
-			0, cos(angle), sin(angle),
-			0, -sin(angle), cos(angle)
+			0, cosf(angle), sinf(angle),
+			0, -sinf(angle), cosf(angle)
 	};
 }
 
 Renderer::Geom::Matrix33f Renderer::Geom::MatrixFactory::rotY(float angle) {
 	return Matrix33f{
-			cos(angle),  0, sin(angle),
+			cosf(angle),  0, sinf(angle),
 			0,           1,      0,
-			-sin(angle), 0, cos(angle)
+			-sinf(angle), 0, cosf(angle)
 	};
 }
 
 Renderer::Geom::Matrix33f Renderer::Geom::MatrixFactory::rotZ(float angle) {
 	return Matrix33f{
-			cos(angle),  sin(angle), 0,
-			-sin(angle), cos(angle), 0,
+			cosf(angle),  sinf(angle), 0,
+			-sinf(angle), cosf(angle), 0,
 			0 ,          0,          1
 	};
 }
@@ -43,10 +43,10 @@ Renderer::Geom::Matrix33f Renderer::Geom::MatrixFactory::scale(const V3<float> &
 Renderer::Geom::Matrix44f Renderer::Geom::MatrixFactory::projection(float left, float right,
 		float top, float bottom, float near, float far) {
 	return Matrix44f{
-		2.0 / (right - left), 0, 0, 0,
-		0, 2.0 / (top - bottom), 0, 0,
+		2.0f / (right - left), 0, 0, 0,
+		0, 2.0f / (top - bottom), 0, 0,
 		(left + right) / (left - right), (bottom + top) / (bottom - top), (far + near) / (far - near), 1,
-		0, 0, -2.0 * near * far / ( far - near ), 0
+		0, 0, -2.0f * near * far / ( far - near ), 0
 
 	};
 }
